Address three post-review gaps in link and participant
Fix the receive-first timer bug in exchangeMappedElements (the
120s backstop kept running while conn.send settled) by delegating
to runReceiveSequence, which resets its inactivity timer on each
received message and clears it immediately on settle.

Extract runReceiveSequence to src/utils/receiveSequence.ts so both
participant.ts and link.ts share the same settled/cleanup/fail/succeed
skeleton, eliminating independent re-implementations that risked
diverging on a future maintenance pass.

Extract StubConnection to test/utils/stubConnection.ts so the identical
class no longer needs to be kept in sync across linkFailures.test.ts and
psiParticipantFailures.test.ts.
