+++ packages/core/src/link.ts
@@ -3,6 +3,8 @@ import * as z from "zod";
 import type { Connection } from "./types";
 import type { PSIParticipant } from "./participant";
 
+import { runReceiveSequence } from "./utils/receiveSequence";
+
 import { getLoggerForVerbosity } from "./utils/logger";
 
 const associationAndIterationArray = z.array(
@@ -236,13 +238,8 @@ export async function linkViaPSI(
   }
 }
 
-// Coarse backstop for a single mapped-element round trip: if the peer neither
-// responds nor surfaces a transport error within this window it is treated as
-// gone. Sized generously so per-message work on large match sets is not cut off.
-const MAPPED_EXCHANGE_TIMEOUT_MS = 120_000;
-
 /** @internal */
-export function exchangeMappedElements(
+export async function exchangeMappedElements(
   id: string,
   conn: Connection,
   log: {
@@ -252,79 +249,31 @@ export function exchangeMappedElements(
   sendFirst: boolean,
   values: IterationMap,
 ): Promise<IterationMap> {
-  return new Promise<IterationMap>((resolve, reject) => {
-    let settled = false;
+  let result: IterationMap | undefined;
 
-    const cleanup = () => {
-      clearTimeout(timer);
-      conn.removeListener("data", onData, undefined, true);
-      conn.removeListener("error", onError, undefined, true);
-    };
-    const fail = (err: unknown) => {
-      if (settled) return;
-      settled = true;
-      cleanup();
-      reject(err instanceof Error ? err : new Error(String(err)));
-    };
-    const succeed = (map: IterationMap) => {
-      if (settled) return;
-      settled = true;
-      cleanup();
-      resolve(map);
-    };
+  await runReceiveSequence(
+    conn,
+    [
+      async (rawData: unknown) => {
+        log.debug(`${id}: received other mapped elements`);
+        result = associationAndIterationArray.parse(rawData);
+        if (!sendFirst) {
+          log.debug(`${id}: sending own mapped elements`);
+          await conn.send(values);
+        }
+      },
+    ],
+    sendFirst
+      ? () => {
+          log.debug(`${id}: sending own mapped elements`);
+          const r = conn.send(values);
+          log.debug(`${id}: waiting for response`);
+          return r;
+        }
+      : undefined,
+  );
 
-    function onError(err: unknown) {
-      fail(err);
-    }
-    function onData(rawData: unknown) {
-      log.debug(`${id}: received other mapped elements`);
-      let parsed: IterationMap;
-      try {
-        parsed = associationAndIterationArray.parse(rawData);
-      } catch (err) {
-        fail(err);
-        return;
-      }
-      if (sendFirst) {
-        succeed(parsed);
-        return;
-      }
-      // Receive-first: send our own values, then resolve once the send settles.
-      log.debug(`${id}: sending own mapped elements`);
-      let sendResult: void | Promise<void>;
-      try {
-        sendResult = conn.send(values);
-      } catch (err) {
-        fail(err);
-        return;
-      }
-      Promise.resolve(sendResult).then(() => succeed(parsed), fail);
-    }
-
-    const timer = setTimeout(
-      () => fail(new Error("PSI mapped-element exchange timed out")),
-      MAPPED_EXCHANGE_TIMEOUT_MS,
-    );
-    conn.once("error", onError);
-    conn.once("data", onData);
-
-    const buffered = conn.takeBufferedError();
-    if (buffered !== undefined) {
-      fail(buffered);
-      return;
-    }
-
-    if (sendFirst) {
-      log.debug(`${id}: sending own mapped elements`);
-      let sendResult: void | Promise<void>;
-      try {
-        sendResult = conn.send(values);
-      } catch (err) {
-        fail(err);
-        return;
-      }
-      Promise.resolve(sendResult).catch(fail);
-      log.debug(`${id}: waiting for response`);
-    }
-  });
+  if (result === undefined)
+    throw new Error("invariant: result was not set by handler");
+  return result;
 }
