+++ packages/core/src/participant.ts
@@ -1,7 +1,5 @@
 import * as z from "zod";
 
-import { EventHandlerQueue } from "./connection/eventHandlerQueue";
-
 import type { AssociationTable, Config, Connection } from "./types";
 
 import type { Client as PSIClient } from "@openmined/psi.js/implementation/client.d.ts";
@@ -9,6 +7,7 @@ import type { PSILibrary } from "@openmined/psi.js/implementation/psi.d.ts";
 import type { Server as PSIServer } from "@openmined/psi.js/implementation/server.d.ts";
 
 import { getLoggerForVerbosity } from "./utils/logger";
+import { runReceiveSequence } from "./utils/receiveSequence";
 import type { ServerSetup } from "@openmined/psi.js/implementation/proto/psi_pb";
 
 const statusCompletedMessage = z.object({
@@ -19,93 +18,6 @@ const numberArrayMessage = z.array(z.number());
 const associationTableMessage = z.array(z.array(z.number()));
 
 const DEFAULT_VERBOSITY = 1;
-
-// Coarse backstop for the PSI set-exchange phase: if no message arrives within
-// this window the peer is treated as gone. The timer is reset on every received
-// message so a long but steadily-progressing exchange is not cut off, and the
-// window is generous so per-message crypto on large datasets does not trip it.
-const PSI_INACTIVITY_TIMEOUT_MS = 120_000;
-
-/**
- * Runs `handlers` in order, one per received `data` event, resolving once the
- * final handler completes. Rejects if `conn` emits `error`, if a handler throws
- * or rejects (e.g. a send to a peer that has dropped), if a transport error was
- * buffered before the phase began, or if no message arrives for
- * {@link PSI_INACTIVITY_TIMEOUT_MS}. `initialSend`, when provided, runs after
- * the listeners are registered (the starter uses it to send its setup message)
- * and its failure rejects the phase. The data/error listeners and the timer are
- * always torn down on settle, so nothing outlives the phase.
- */
-function runReceiveSequence(
-  conn: Connection,
-  handlers: Array<(rawData: unknown) => void | Promise<void>>,
-  initialSend?: () => void | Promise<void>,
-): Promise<void> {
-  return new Promise<void>((resolve, reject) => {
-    let settled = false;
-    let timer: ReturnType<typeof setTimeout>;
-
-    const cleanup = () => {
-      clearTimeout(timer);
-      conn.removeListener("data", dataListener);
-      conn.removeListener("error", onError, undefined, true);
-    };
-    const fail = (err: unknown) => {
-      if (settled) return;
-      settled = true;
-      cleanup();
-      reject(err instanceof Error ? err : new Error(String(err)));
-    };
-    const succeed = () => {
-      if (settled) return;
-      settled = true;
-      cleanup();
-      resolve();
-    };
-    const onError = (err: unknown) => fail(err);
-    const resetTimer = () => {
-      clearTimeout(timer);
-      timer = setTimeout(
-        () => fail(new Error("PSI exchange timed out")),
-        PSI_INACTIVITY_TIMEOUT_MS,
-      );
-    };
-
-    const lastIndex = handlers.length - 1;
-    const queue = new EventHandlerQueue(
-      handlers.map((handler, index) => async (rawData: unknown) => {
-        await handler(rawData);
-        if (index === lastIndex) succeed();
-      }),
-      fail,
-    );
-    const dataListener = (rawData: unknown) => {
-      resetTimer();
-      queue.handleEvent(rawData);
-    };
-
-    resetTimer();
-    conn.once("error", onError);
-    conn.on("data", dataListener);
-
-    const buffered = conn.takeBufferedError();
-    if (buffered !== undefined) {
-      fail(buffered);
-      return;
-    }
-
-    if (initialSend !== undefined) {
-      let sendResult: void | Promise<void>;
-      try {
-        sendResult = initialSend();
-      } catch (err) {
-        fail(err);
-        return;
-      }
-      Promise.resolve(sendResult).catch(fail);
-    }
-  });
-}
 
 export enum ProcessState {
   BeforeStart,
